test_runner externalized
