+++ smart-contract/program/tests/common/mod.rs
@@ -1,1 +1,2 @@
 pub mod utils;
+pub mod test_runner;
